pyproject.toml
[tool.poetry]
name = "habit_tracker"
version = "0.1.0"
description = ""
authors = []

[tool.poetry.dependencies]
python = "^3.7"
django = "^3.0.6"
django-extensions = "^2.2.9"
django-debug-toolbar = "^2.2"
werkzeug = "^1.0.1"
django-environ = "^0.4.5"
django-registration-redux = "^2.8"
LemmInflect = "^0.2.1"

[tool.poetry.dev-dependencies]
pylint = "^2.5.3"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
